+++ src/web/extension/symbol.ts
@@ -5,8 +5,9 @@ import {
 } from 'greyscript-core';
 import Monaco from 'monaco-editor/esm/vs/editor/editor.api.js';
 
-import ASTStringify from './helper/ast-stringify.js';
+import transformASTToString from './helper/ast-stringify.js';
 import documentParseQueue from './helper/model-manager.js';
+import { removeContextPrefixInNamespace } from './helper/utils.js';
 
 const findAllAssignments = (
   monaco: typeof Monaco,
@@ -20,7 +21,9 @@ const findAllAssignments = (
   for (const item of scopes) {
     for (const assignmentItem of item.assignments) {
       const assignment = assignmentItem as ASTAssignmentStatement;
-      const current = ASTStringify(assignment.variable);
+      const current = removeContextPrefixInNamespace(
+        transformASTToString(assignment.variable)
+      );
 
       if (!isValid(current)) {
         continue;
@@ -36,6 +39,7 @@ const findAllAssignments = (
         assignment.variable.end.line,
         assignment.variable.end.character
       );
+
       result.push({
         name: current,
         detail: current,